src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["KarimdzanxSifei"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

numpy = ">=1.24"
pandas = ">=2.0"
scipy = ">=1.10"
pyyaml = ">=6.0"

scikit-learn = ">=1.3"
tqdm = ">=4.66"

implicit = ">=0.7.2"
lightfm = ">=1.17"

polars = ">=0.20"
pyarrow = ">=14.0"

datasets = ">=2.18"
transformers = ">=4.41"
torch = ">=2.1"

vllm = ">=0.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
